requirements.txt
PyQt6
adbutils
qt-material
pyinstaller
pyaxmlparser
fpdf2
odfpy
